Major Update #3 - JS for "Code"
Added the javascript for the main menu's "Code" button.
It is likely still very buggy and one of the messier pieces I've ever
written;  but it is 2:30 and I want to keep to this nightly commit
schedule before bed.

Side note: Shawn is very proud of his first ever JS, despite the
clunkiness of it.